requirements.txt
cryptography>=46.0.3
email-validator>=2.3.0
fastapi>=0.128.0
pymysql>=1.1.2
python-dotenv>=1.2.1
python-jose[cryptography]>=3.3.0
python-multipart>=0.0.9
bcrypt>=4.0.0
sqlalchemy>=2.0.45
uvicorn
